requirements.txt
dotenv~=0.9.9 

## discord
    #audioop-lts==0.2.1 
    discord.py~=2.5.2 
    discord.py[voice]

## API
    httpx~=0.28.1
    websockets~=15.0.1

## Voice Caption
    discord-ext-voice-recv
    soxr
    opuslib~=2.0.0
    audresample~=1.3.4
    #ffmpeg
    #sounddevice
